src/requirements.txt


Django==2.1.7
django-heroku==0.3.1
djangorestframework==3.9.1
gunicorn==19.9.0
Pillow==5.4.1
whitenoise

